requirements.txt
pandas
numpy
scikit-learn
lightgbm
joblib
fastapi
uvicorn
pydantic
python-dotenv
requests
